+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/objects/FPropertyTag.kt
@@ -3,9 +3,9 @@ package me.fungames.jfortniteparse.ue4.assets.objects
 import me.fungames.jfortniteparse.exceptions.ParserException
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
-import me.fungames.jfortniteparse.ue4.assets.util.FName
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
 import me.fungames.jfortniteparse.ue4.objects.core.misc.FGuid
+import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
 
 @ExperimentalUnsignedTypes
 class FPropertyTag : UClass {
