+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/objects/FStructFallback.kt
@@ -3,8 +3,8 @@ package me.fungames.jfortniteparse.ue4.assets.objects
 import me.fungames.jfortniteparse.exceptions.ParserException
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
-import me.fungames.jfortniteparse.ue4.assets.util.FName
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
+import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
 
 @ExperimentalUnsignedTypes
 class FStructFallback : UClass {
