+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/objects/meshes/FStaticMaterial.kt
@@ -3,7 +3,7 @@ package me.fungames.jfortniteparse.ue4.assets.objects.meshes
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.exports.mats.UMaterialInterface
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
-import me.fungames.jfortniteparse.ue4.assets.util.FName
+import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
 import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FPackageIndex
 import me.fungames.jfortniteparse.ue4.versions.FRenderingObjectVersion
