+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/writer/FAssetArchiveWriter.kt
@@ -1,8 +1,8 @@
 package me.fungames.jfortniteparse.ue4.assets.writer
 
 import me.fungames.jfortniteparse.exceptions.ParserException
-import me.fungames.jfortniteparse.ue4.assets.util.FName
 import me.fungames.jfortniteparse.ue4.assets.util.PayloadType
+import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
 import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FNameEntry
 import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FObjectExport
 import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FObjectImport
