+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/objects/coreuobject/uobject/FName.kt
@@ -1,6 +1,4 @@
-package me.fungames.jfortniteparse.ue4.assets.util
-
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FNameEntry
+package me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject
 
 @Suppress("EXPERIMENTAL_API_USAGE")
 open class FName(private val nameMap: List<FNameEntry>, val index: Int, val extraIndex: Int) {
