+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/objects/coreuobject/uobject/SoftObjectPath.kt
@@ -2,7 +2,6 @@ package me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject
 
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
-import me.fungames.jfortniteparse.ue4.assets.util.FName
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
 
 /**
