+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/objects/engine/animation/SmartName.kt
@@ -2,8 +2,8 @@ package me.fungames.jfortniteparse.ue4.objects.engine.animation
 
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
-import me.fungames.jfortniteparse.ue4.assets.util.FName
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
+import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
 
 @ExperimentalUnsignedTypes
 class FSmartName : UClass {
