+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/pak/PakFileReader.kt
@@ -16,12 +16,16 @@ import me.fungames.jfortniteparse.ue4.reader.FByteArchive
 import me.fungames.jfortniteparse.ue4.versions.GAME_UE4
 import me.fungames.jfortniteparse.ue4.versions.GAME_UE4_GET_AR_VER
 import me.fungames.jfortniteparse.ue4.versions.LATEST_SUPPORTED_UE4_VERSION
-import me.fungames.jfortniteparse.util.*
-import me.fungames.kotlinPointers.BytePointer
+import me.fungames.jfortniteparse.util.INDEX_NONE
+import me.fungames.jfortniteparse.util.parseHexBinary
+import me.fungames.jfortniteparse.util.printAesKey
 import mu.KotlinLogging
 import java.io.File
 import java.io.RandomAccessFile
 import java.nio.ByteBuffer
+import kotlin.collections.component1
+import kotlin.collections.component2
+import kotlin.collections.set
 import kotlin.math.min
