+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/registry/reader/FNameTableArchive.kt
@@ -1,7 +1,7 @@
 package me.fungames.jfortniteparse.ue4.registry.reader
 
 import me.fungames.jfortniteparse.exceptions.ParserException
-import me.fungames.jfortniteparse.ue4.assets.util.FName
+import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
 import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FNameEntry
 import me.fungames.jfortniteparse.ue4.reader.FArchive
 import kotlin.math.min
